+++ appboy-segment-integration/src/androidTest/java/com/segment/analytics/android/integrations/appboy/AppboyAndroidTest.java
@@ -5,10 +5,8 @@ import android.support.test.runner.AndroidJUnit4;
 import com.appboy.Appboy;
 import com.appboy.configuration.AppboyConfig;
 import com.segment.analytics.Analytics;
-import com.segment.analytics.Traits;
 import com.segment.analytics.integrations.IdentifyPayload;
 import com.segment.analytics.integrations.Logger;
-import com.segment.analytics.test.IdentifyPayloadBuilder;
 
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -30,12 +28,14 @@ public class AppboyAndroidTest {
   @Test
   public void testIdentifyCallsChangeUser() {
     String testUserId = "testUser" + System.currentTimeMillis();
-    Traits traits = createTraits(testUserId);
-    IdentifyPayload identifyPayload = new IdentifyPayloadBuilder().traits(traits).build();
-    Logger logger = Logger.with(Analytics.LogLevel.DEBUG);
-    AppboyIntegration integration = new AppboyIntegration(Appboy.getInstance(getContext()), "foo", logger, true);
-    integration.identify(identifyPayload);
+    IdentifyPayload identifyPayload = new IdentifyPayload
+      .Builder()
+      .userId(testUserId)
+      .traits(createTraits(testUserId))
+      .build();
+    AppboyIntegration integration = new AppboyIntegration(Appboy.getInstance(getContext()), "token", Logger.with(Analytics.LogLevel.DEBUG), true);
 
+    integration.identify(identifyPayload);
     assertEquals(testUserId, Appboy.getInstance(getContext()).getCurrentUser().getUserId());
   }
 }